Fix deadlock
From the docs
You should be careful to avoid deadlocks. Since pipes are fixed length buffers,
Open3.popen3(“prog”) {|i, o, e, t| o.read } deadlocks if the program generates
too much output on stderr. You should read stdout and stderr simultaneously
(using threads or IO.select)

This was triggering in case of haskell as stack was outputting a lot of
logs potentailly overwhelming the buffers.

This also has an added benefit of streaming logs which was not happening
before.